+++ changelog.md
@@ -2,5 +2,6 @@
 
 - Added keybind to toggle if pings are shared
 - Added keybind to clear all waypoints of the current dimension
+- Added ability to share pings with a custom command
 - Changed waypoint toggling: now only toggles between waypoints of the current dimension
 - Changed ping sound to be always in audible range